% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.expand_and_normalize_path}
\alias{.expand_and_normalize_path}
\title{Expand and normalize a file path.}
\usage{
.expand_and_normalize_path(p, add_trailing_slash = FALSE)
}
\arguments{
\item{p}{Character. The path to expand and normalize.}

\item{add_trailing_slash}{Logical. If `TRUE`, a trailing slash is appended to
the normalized path if it does not already have one. This is useful if the
path is known to be a directory. Defaults to `FALSE`.}
}
\value{
Character. The normalized path, or `NULL` if the path is invalid or
  does not exist.
}
\description{
Symbolic links and ".." will be followed and expanded.
}
\keyword{internal}
