% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.file_relative_to_path}
\alias{.file_relative_to_path}
\title{Format the file path to be relative to relative_to_path}
\usage{
.file_relative_to_path(file, relative_to_path)
}
\arguments{
\item{file}{The file path to format.}

\item{relative_to_path}{The path that we want the file to be
relative to.}
}
\value{
The formatted file path. If either `file` or `relative_to_path`
  do not exist, or if `file` is not a descendant of `relative_to_path` then
  simply the basename of `file` is returned.
}
\description{
This is generally for informational purposes to report to the user. It is
meant to hide the full paths of files on the system from a user so that
attackers cannot gather information about the system's directory structure.
Usually, the relative_to_path parameter would be the sandbox path.
}
\keyword{internal}
