% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.get_file}
\alias{.get_file}
\title{Get File from Model Cache}
\usage{
.get_file(mod, file)
}
\arguments{
\item{mod}{Model object}

\item{file}{Path to file to retrieve}
}
\value{
Data frame from the file cache
}
\description{
Internal function that retrieves a previously loaded file from the
model's cache.
}
\keyword{internal}
