% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.get_string_parts}
\alias{.get_string_parts}
\title{Split and trim string parts}
\usage{
.get_string_parts(s, split = ";")
}
\arguments{
\item{s}{Character string to split}

\item{split}{Character string to use as delimiter (default: ";")}
}
\value{
Character vector of trimmed string parts
}
\description{
Split and trim string parts
}
\keyword{internal}
