% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.is_file_descendant_of}
\alias{.is_file_descendant_of}
\title{Check if a file is a descendant of a directory, and that both the file and
directory exist.}
\usage{
.is_file_descendant_of(file, top_level_directory)
}
\arguments{
\item{file}{Character. The path to the file to check.}

\item{top_level_directory}{Character. The path to the directory that `file`
should be a descendant of.}
}
\value{
Logical. `TRUE` if `file` is a descendant of `top_level_directory`,
  `FALSE` otherwise. Returns `FALSE` if either `file` or
  `top_level_directory` do not exist on the file system.
}
\description{
Symbolic links and ".." will be followed and expanded.
}
\keyword{internal}
