% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-rcs.R
\name{.run_step_rcs}
\alias{.run_step_rcs}
\title{Run Restricted Cubic Spline Step}
\usage{
.run_step_rcs(mod, file)
}
\arguments{
\item{mod}{Model object}

\item{file}{Path to RCS step specification file}
}
\value{
Updated model object with RCS variables added to data
}
\description{
Creates restricted cubic spline (RCS) transformations using
specified knot positions. Implements the 'rcs' transformation step
from the Model Parameters pipeline.
}
\keyword{internal}
