% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.get_unused_column}
\alias{.get_unused_column}
\title{Find an unused column name in the data}
\usage{
.get_unused_column(data, column_prefix)
}
\arguments{
\item{data}{Data to check for existing column names.}

\item{column_prefix}{Character prefix for the column name}
}
\value{
Character string of an unused column name
  (e.g., "prefix1", "prefix2")
}
\description{
Generates a unique column name by appending an integer to a prefix.
Iteratively checks for column_prefix1, column_prefix2, etc. until
finding a name that doesn't exist in the data (eg. a dataframe).
}
\keyword{internal}
