% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-dummy.R
\name{.run_step_dummy}
\alias{.run_step_dummy}
\title{Run Dummy Coding Step}
\usage{
.run_step_dummy(mod, file)
}
\arguments{
\item{mod}{Model object containing input data in \code{mod$data}}

\item{file}{Path to dummy step specification file}
}
\value{
A list containing: \code{mod} (the updated model object with
  dummy variables added to \code{mod$data}), and \code{output_columns}
  (character vector of output columns of this step)
}
\description{
Creates dummy variables for categorical values, setting 1 when the
original variable equals the specified category value, 0 otherwise.
Implements the 'dummy' transformation step from the Model
Parameters pipeline.
}
\keyword{internal}
