% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline_utils.R
\name{.verify_columns}
\alias{.verify_columns}
\title{Verify required columns exist in a dataframe}
\usage{
.verify_columns(df, columns, data_description, file = NULL)
}
\arguments{
\item{df}{Data frame to validate}

\item{columns}{Character vector of required column names}

\item{data_description}{Description of the data being validated to include
in the error message.}

\item{file}{Optional file path to include in error message (to indicate
where the data originated from). For example: "model steps file" or
"model steps data".}
}
\description{
Checks that all required columns are present in the dataframe and stops
with an informative error message if any are missing.
}
\keyword{internal}
