% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline.R
\name{run_model_pipeline}
\alias{run_model_pipeline}
\title{Run Model Parameters Pipeline}
\usage{
run_model_pipeline(
  root_dir = NULL,
  data = NULL,
  model_export = NULL,
  variables = NULL,
  model_steps = NULL,
  existing_mod = NULL
)
}
\arguments{
\item{root_dir}{Root directory for model parameter files. This is only
used when loading the step files found in the model steps specification.
The other parameters to this function must include the full path
location (ie. data, model_export, variables, and model_steps)}

\item{data}{Input data (file path or data frame)}

\item{model_export}{Model export specification (file path or data frame)}

\item{variables}{Variables specification (file path or data frame)}

\item{model_steps}{Model steps specification (file path or data frame)}

\item{existing_mod}{Existing model object to reuse (for performance
optimization)}
}
\value{
Model object with transformed data in `$df` component
}
\description{
Executes the complete model parameters transformation pipeline, applying
sequential steps (center, dummy, interaction, rcs, logistic) to input
data.
}
