% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_pipeline.R
\name{run_model_pipeline}
\alias{run_model_pipeline}
\title{Run Model Pipeline}
\usage{
run_model_pipeline(mod, x, mode = "output")
}
\arguments{
\item{mod}{A model object created by \code{\link{prepare_model_pipeline}}.}

\item{x}{Either a file path (character) to a CSV file containing the
input data, or a data frame. The data must contain all columns specified
as predictors in the variables file.}

\item{mode}{A character string specifying what data to return. Can be one
of:
\itemize{
   \item "output": Only return the final output of the model. These are the
     values of all variables calculated in the final step found in the
     model steps file.
   \item "full": Return all data, which includes the input data, all
     intermediate variables, and the final output of the model.
}
Default is "output".}
}
\value{
A model object created from a call to
  \code{\link{prepare_model_pipeline}}.
}
\description{
Executes the transformation pipeline on input data and retrieve the output.
Applies each transformation step defined in the model steps file in
sequence, modifying the data accordingly.
}
\examples{
\dontrun{
# Prepare and run pipeline
mod <- prepare_model_pipeline("path/to/model-export.csv")
output <- run_model_pipeline(mod, x = "path/to/input-data.csv")
head(output)

# Run on data frame
input_data <- read.csv("path/to/data.csv")
mod <- run_model_pipeline(mod, x = input_data)
}

}
\seealso{
\code{\link{prepare_model_pipeline}} to prepare the model object
}
